test(cli): silence logs for deterministic output
