+++ deno.json
@@ -14,6 +14,7 @@
     "@std/log": "https://deno.land/std@0.224.0/log/mod.ts",
     "@std/path": "https://deno.land/std@0.224.0/path/mod.ts",
     "@std/fs": "https://deno.land/std@0.224.0/fs/mod.ts",
+    "@std/flags": "https://deno.land/std@0.224.0/flags/mod.ts",
     "zod": "https://deno.land/x/zod@v3.22.4/mod.ts",
     "pg": "https://deno.land/x/postgres@v0.17.0/mod.ts",
     "sqlite3": "https://deno.land/x/sqlite3@v0.9.2/mod.ts"
